Add support for a database proxy host
If you are connecting securely to a database chances are you have locked
down the addresses that can access it. In some cases such as GCP the
recommended way of connecting to a database instance is through a proxy
or SSH tunnel (CloudSQL Auth Proxy in GCP's case).

If set, `dbProxyConnectionHost` will replace the `host` value of the
postgres connection string. This allows you to specify a proxy address
(probably 127.0.0.1) to connect to the database initially, whilst
passing a different connection to Hasura for use when connecting from
Hasura Cloud.